Added a way to edit volitility, not working rn #30
when the graph is started, the last entry is always wildly different from the rest. this doesn't seem to happend when StartGraph is called from the Start button.

Curently the update is not working as expected. When the graph is updated, the first entry is very different. but the following the entrys look stable like it's following the previous volatility value.